Consolidate metric dashboard E2E coverage
